refactor: move Rocket.Chat credentials to adapter
