+++ src/main/java/de/caritas/cob/userservice/api/model/EnquiryData.java
@@ -1,6 +1,6 @@
 package de.caritas.cob.userservice.api.model;
 
-import de.caritas.cob.userservice.api.container.RocketChatCredentials;
+import de.caritas.cob.userservice.api.adapters.rocketchat.RocketChatCredentials;
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
 import lombok.Getter;
